feat: 私有化更新定制内容回合
1. 增加对更新策略的适配；
2. 将更新开始、更新结束时间上报给私有化更新平台；
3. 更新完成后若更新结果上报失败，增加重启后重试和定时重试；

Log: 私有化更新定制内容回合
Task: https://pms.uniontech.com/task-view-360227.html
Influence: 私有化更新定制内容回合

+++ src/lastore-daemon/manager_download.go
@@ -75,7 +75,7 @@ func (m *Manager) prepareDistUpgrade(sender dbus.Sender, origin system.UpdateTyp
 			string(system.ErrorInsufficientSpace),
 			"You don't have enough free space to download",
 		}
-		msg := fmt.Sprintf(gettext.Tr("Downloading updates failed. Please free up %n GB disk space first."), needDownloadSize/(1000*1000*1000))
+		msg := fmt.Sprintf(gettext.Tr("Downloading updates failed. Please free up %g GB disk space first."), needDownloadSize/(1000*1000*1000))
 		go m.sendNotify(updateNotifyShowOptional, 0, "preferences-system", "", msg, nil, nil, system.NotifyExpireTimeoutNoHide)
 		logger.Warning(dbusError.Detail)
 		errStr, _ := json.Marshal(dbusError)
@@ -131,15 +131,17 @@ func (m *Manager) prepareDistUpgrade(sender dbus.Sender, origin system.UpdateTyp
 			m.PropsMu.Lock()
 			m.PropsMu.Unlock()
 			m.statusManager.SetUpdateStatus(mode, system.IsDownloading)
-			sendDownloadingOnce.Do(func() {
-				msg := gettext.Tr("New version available! Downloading...")
-				action := []string{
-					"view",
-					gettext.Tr("View"),
-				}
-				hints := map[string]dbus.Variant{"x-deepin-action-view": dbus.MakeVariant("dde-control-center,-m,update")}
-				go m.sendNotify(updateNotifyShowOptional, 0, "preferences-system", "", msg, action, hints, system.NotifyExpireTimeoutDefault)
-			})
+			if !m.updatePlatform.UpdateNowForce { // 立即更新则不发通知
+				sendDownloadingOnce.Do(func() {
+					msg := gettext.Tr("New version available! Downloading...")
+					action := []string{
+						"view",
+						gettext.Tr("View"),
+					}
+					hints := map[string]dbus.Variant{"x-deepin-action-view": dbus.MakeVariant("dde-control-center,-m,update")}
+					go m.sendNotify(updateNotifyShowOptional, 0, "preferences-system", "", msg, action, hints, system.NotifyExpireTimeoutDefault)
+				})
+			}
 			return
 		}
 		j.setPreHooks(map[string]func() error{
@@ -167,7 +169,7 @@ func (m *Manager) prepareDistUpgrade(sender dbus.Sender, origin system.UpdateTyp
 							logger.Warning(err)
 							size = needDownloadSize
 						}
-						msg = fmt.Sprintf(gettext.Tr("Downloading updates failed. Please free up %n GB disk space first."), size/(1000*1000*1000))
+						msg = fmt.Sprintf(gettext.Tr("Downloading updates failed. Please free up %g GB disk space first."), size/(1000*1000*1000))
 						go m.sendNotify(updateNotifyShowOptional, 0, "preferences-system", "", msg, nil, nil, system.NotifyExpireTimeoutDefault)
 					} else if strings.Contains(errorContent.ErrType, string(system.ErrorDamagePackage)) {
 						// 下载更新失败，需要apt-get clean后重新下载
@@ -195,24 +197,26 @@ func (m *Manager) prepareDistUpgrade(sender dbus.Sender, origin system.UpdateTyp
 			string(system.SucceedStatus): func() error {
 				m.statusManager.SetUpdateStatus(j.updateTyp, system.CanUpgrade)
 				if j.next == nil {
-					msg := gettext.Tr("Downloading completed. You can install updates when shutdown or reboot.")
-					action := []string{
-						"updateNow",
-						gettext.Tr("Update Now"),
-						"ignore",
-						gettext.Tr("Dismiss"),
-					}
-					hints := map[string]dbus.Variant{"x-deepin-action-updateNow": dbus.MakeVariant("dbus-send,--session,--print-reply,--dest=com.deepin.dde.shutdownFront,/com/deepin/dde/shutdownFront,com.deepin.dde.shutdownFront.Show")}
 					go func() {
 						m.inhibitAutoQuitCountAdd()
 						defer m.inhibitAutoQuitCountSub()
-						m.sendNotify(updateNotifyShowOptional, 0, "preferences-system", "", msg, action, hints, system.NotifyExpireTimeoutDefault)
+						if !m.updatePlatform.UpdateNowForce {
+							msg := gettext.Tr("Downloading completed. You can install updates when shutdown or reboot.")
+							action := []string{
+								"updateNow",
+								gettext.Tr("Update Now"),
+								"ignore",
+								gettext.Tr("Dismiss"),
+							}
+							hints := map[string]dbus.Variant{"x-deepin-action-updateNow": dbus.MakeVariant("dbus-send,--session,--print-reply,--dest=com.deepin.dde.shutdownFront,/com/deepin/dde/shutdownFront,com.deepin.dde.shutdownFront.Show")}
+							m.sendNotify(updateNotifyShowOptional, 0, "preferences-system", "", msg, action, hints, system.NotifyExpireTimeoutDefault)
+						}
 						m.reportLog(downloadStatusReport, true, "")
 					}()
 
 					if m.updatePlatform.UpdateNowForce {
 						m.inhibitAutoQuitCountAdd()
-						_, err := m.distUpgradePartly(dbus.Sender(m.service.Conn().Names()[0]), j.updateTyp, true)
+						_, err := m.distUpgradePartly(dbus.Sender(m.service.Conn().Names()[0]), mode, true)
 						if err != nil {
 							logger.Error("failed to dist-upgrade:", err)
 						}
